requirements.txt
duckdb
pandas
numpy
python-dotenv
google-genai
